+++ include/rapidjson/error/en.h
@@ -120,6 +120,9 @@ inline const RAPIDJSON_ERROR_CHARTYPE* GetValidateError_En(ValidateErrorCode val
         case kNormalizeErrorCircularAlias:              return RAPIDJSON_ERROR_STRING("Property has circular aliases.");
         case kNormalizeErrorConflictingAliases:         return RAPIDJSON_ERROR_STRING("Property has conflicting aliases.");
         case kNormalizeErrorMergeConflict:              return RAPIDJSON_ERROR_STRING("Property has conflicting normalized documents.");
+	// Warnings
+        case kValidateWarnings:                         return RAPIDJSON_ERROR_STRING("One or more validation warnings have occurred");
+        case kDeprecatedWarning:                        return RAPIDJSON_ERROR_STRING("Property is being deprecated.");
 #endif // RAPIDJSON_YGGDRASIL
         default:                                        return RAPIDJSON_ERROR_STRING("Unknown error.");
     }